chore: replace vector feature map with npm version
