--- public/frontend/src/components/StyledVectorFeatureMap/hooks/index.ts
@@ -1,5 +1,0 @@
-export * from './useMapInitialization';
-export * from './useMapBaseLayers';
-export * from './useAddVectorLayerToMap';
-export * from './useGetMapStyles';
-export * from './useOpenLayersTracking';
