--- public/frontend/src/components/StyledVectorFeatureMap/index.ts
@@ -1,1 +1,0 @@
-export * from './StyledVectorFeatureMap';
