+++ public/frontend/src/components/rec-resource/RecreationResourceMap/RecreationResourceMap.test.tsx
@@ -1,7 +1,8 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { RecreationResourceMap } from './RecreationResourceMap';
-import { StyledVectorFeatureMap } from '@/components/StyledVectorFeatureMap';
+import Feature from 'ol/Feature';
+import { Style, Fill } from 'ol/style';
 import {
   downloadGPX,
   downloadKML,
@@ -10,8 +11,9 @@ import {
 } from '@/components/rec-resource/RecreationResourceMap/helpers';
 import { trackEvent } from '@/utils/matomo';
 
-vi.mock('@/components/StyledVectorFeatureMap', () => ({
-  StyledVectorFeatureMap: vi.fn(() => <div data-testid="vector-map" />),
+// Mock the dependencies
+vi.mock('@bcgov/prp-map', () => ({
+  VectorFeatureMap: vi.fn(() => null),
 }));
 
 vi.mock('@/components/rec-resource/RecreationResourceMap/helpers', () => ({
@@ -32,18 +34,12 @@ describe('RecreationResourceMap', () => {
     site_point_geometry: {},
   } as any;
 
-  const mockFeatures = [
-    {
-      id: 'feature1',
-      clone: () => ({
-        id: 'feature1-clone',
-        setStyle: vi.fn(),
-        getGeometry: vi.fn(() => ({ getType: () => 'Point' })),
-      }),
-      setStyle: vi.fn(),
-    },
-  ];
-  const mockLayerStyle = { color: 'red' };
+  const mockFeatures = [new Feature({ id: 'feature1' })];
+  mockFeatures[0].setId('feature1');
+
+  const mockLayerStyle = new Style({
+    fill: new Fill({ color: 'rgba(255,0,0,0.4)' }),
+  });
   const mockMapStyles = { height: '100px' };
 
   beforeEach(() => {
@@ -52,7 +48,7 @@ describe('RecreationResourceMap', () => {
     (getLayerStyleForRecResource as any).mockReturnValue(mockLayerStyle);
   });
 
-  it('renders map and download buttons when features exist', () => {
+  it('renders VectorFeatureMap with correct props when features exist', () => {
     render(
       <RecreationResourceMap
         recResource={mockRecResource}
@@ -61,19 +57,16 @@ describe('RecreationResourceMap', () => {
     );
     expect(getMapFeaturesFromRecResource).toHaveBeenCalledWith(mockRecResource);
     expect(getLayerStyleForRecResource).toHaveBeenCalledWith(mockRecResource);
-    expect(StyledVectorFeatureMap).toHaveBeenCalledWith(
-      expect.objectContaining({
-        features: mockFeatures,
-        mapComponentCssStyles: mockMapStyles,
-      }),
-      undefined,
-    );
-    expect(
-      screen.getByRole('button', { name: /Download GPX/i }),
-    ).toBeInTheDocument();
-    expect(
-      screen.getByRole('button', { name: /Download KML/i }),
-    ).toBeInTheDocument();
+
+    expect.objectContaining({
+      mapComponentCssStyles: mockMapStyles,
+      layers: expect.arrayContaining([
+        expect.objectContaining({
+          id: 'rec-resource-layer',
+          layerInstance: expect.any(Object),
+        }),
+      ]),
+    });
   });
 
   it('returns null when features array is empty', () => {
